requirements.txt
Telethon~=1.23.0
python-dotenv~=0.19.0